2 Mpu6050 Reading Seperately
Connected two separate gyroscopes to the Gauntlet. Shoulder is MPU and Wrist is MPU0, both can accurately read and print rotation data to the serial monitor for viewing. Will be changing the code to accurately display the rotation in positive values as well as finding rotation between them, and creating the Data packet array.